added/ Accordions Component and moved logic from Faq to Accordions
